Remove duplicate data builder of identity server module.
Resolve #
